updated main.js. Renamed DessertOptions.js to desserts.js
